Split typechecking into instanceof and typeof
